+++ src/utils/interfaces.ts
@@ -1,0 +1,6 @@
+export interface Article {
+	title: string;
+	url: string;
+	languageCode: string;
+	description: string | null;
+}
