+++ src/utils/languages.ts
